Move alleged offences test to its own subdirectory
